requirements.txt
# Core dependencies
# python>=3.10

# Document parsing
pymupdf>=1.23.0
trafilatura>=1.6.0
pytesseract>=0.3.10  # Optional OCR

# NLP and embeddings
sentence-transformers>=2.2.0
spacy>=3.7.0
transformers>=4.35.0

# Vector storage
faiss-cpu>=1.7.4
chromadb>=0.4.0

# LLM frameworks (choose one or both)
langchain>=0.1.0
langchain-community>=0.0.10
# autogen-agentchat>=0.2.0  # Alternative

# LLM API (optional, for enhanced features)
openai>=1.0.0  # For OpenAI, Zhipu, DeepSeek and compatible APIs

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Frontend
streamlit>=1.28.0
plotly>=5.18.0  # For interactive charts
# gradio>=4.0.0  # Alternative

# Database
# neo4j>=5.14.0  # Optional for KG visualization

# Utilities
pydantic>=2.5.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0

# Evaluation
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

